Add a new strict parameter to get_nodes() function
Sometimes, the openstreetmap server is inconsistent in that way that
the list of node ids of a `Way` contains nodes that actually do not exist.

In that case, no nodes of a `Way` will be retrieved even if there are
valid nodes.

You can now set the strict_mode to False so that invalid nodes will be
removed from a Way.

The default behaviour still is to try to retrieve the missing nodes and
if the node does not exist, a DataIncomplete exception is thrown.